% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Shiny App}
\usage{
run_app()
}
\value{
A running Shiny app
}
\description{
This function starts the Shiny app included in the package.
}
\examples{
if (interactive()) {
  run_app()
}
}
